+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 
 * Recycled code.
 * Fixed compatibility with the majority of addons.
+* Added filter "With multiple models" and a button to control the models.